fix: provide port base url as default and correct test action (#5)
Our integrations were broken because:
- A required parameter was added for the integration to work on the
backend, but wasn't set as required in the GitHub action inputs:
`pydantic.error_wrappers.ValidationError: 1 validation error for
IntegrationConfiguration
port -> base_url`

- The required parameter `https://app.getport.io` should be
`https://api.getport.io`

I left it as a default to `https://api.getport.io` which makes sense to
me but of course feel free to change. Adding an extra input for the API
URL seems like it's not needed for the user if it's the same for
everything.